User::create return int

+++ admin/users.php
@@ -167,7 +167,7 @@ switch ($_REQUEST['action']) {
         }
 
         /* Attempt to create the user */
-        $user_id = (int) User::create($username, $fullname, $email, $website, $pass1, (string) $access, $state, $city);
+        $user_id = User::create($username, $fullname, $email, $website, $pass1, (string) $access, $state, $city);
         if ($user_id < 1) {
             AmpError::add('general', T_("The new User was not created"));
         }
